% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate a \code{wsre} object}
\usage{
evaluate(wsre_obj, x_nu, x_de)
}
\arguments{
\item{wsre_obj}{\code{wsre} object: From \code{\link{wsre}}}

\item{x_nu}{Numeric Vector: vector of numerator locations to evaluate the
estimate at.}

\item{x_de}{Numeric Vector: vector of denominator locations to evaluate the
estimate at.}
}
\value{
Numeric Vector: self ratio values
}
\description{
Evaluate a weighted self ratio estimate at a set of numerator and denominator
points
}
\details{
Note that \code{\link[base]{expand.grid}} gets called on \code{x_nu} and 
\code{x_de}.
}
