% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\docType{data}
\name{null_wsre_data}
\alias{null_wsre_data}
\title{Null wsre values}
\format{
An object of class \code{list} of length 4.
}
\usage{
null_wsre_data
}
\description{
Null wsre parameter values, for when you want to use the included stan models
but not do wsre.
}
\keyword{datasets}
