% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsre-package.R
\docType{package}
\name{wsre-package}
\alias{wsre-package}
\title{The 'wsre' package.}
\description{
Weighted Self-Density Ratio estimation
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.1. https://mc-stan.org
}
