% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{wsre}
\alias{wsre}
\title{Weighted Self Ratio Estimation}
\usage{
wsre(model_name = c("normal", "binom"), wf_mean = c(-3, 3, 5),
  wf_pars = list(wf_sd = 2, wf_exponent = 1, target_dimension = 1),
  n_mcmc_samples = 5000, stan_control_params = list(adapt_delta = 0.95,
  max_treedepth = 12), flog_threshold = futile.logger::INFO)
}
\arguments{
\item{model_name}{String: "normal" or "binom" at the moment. Must be the name
of a \code{.stan} file in the \code{stan_files} directory.}

\item{wf_mean}{Numeric Vector: vector of means of weighting functions. May be
computed automatically in the future.}

\item{wf_pars}{Named List: see default for structure; other parameters for
the weighting function}

\item{n_mcmc_samples}{Integer: number of MCMC samples to draw from each of 
the targets.}

\item{stan_control_params}{Named List: see \code{control} section of
\code{\link[rstan]{stan}}.}

\item{flog_threshold}{Special: See 
\code{\link[futile.logger]{futile.logger-package}}. Set to 
\code{futile.logger::TRACE} to see informational messages about what is going
on.}
}
\value{
A \code{wsre} object for saving / use in other functions
}
\description{
Estimates the self ratio of a given model using a number of weighting 
functions.
}
