% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate a \code{wsre} object}
\usage{
evaluate(wsre_obj, x_nu, x_de, mc_cores = 1)
}
\arguments{
\item{wsre_obj}{\code{wsre} object: From \code{\link{wsre}}}

\item{x_nu}{Numeric Vector: vector of length one or more; numerator
location to evaluate the estimate at.}

\item{x_de}{Numeric Vector: vector of length one or more; denominator
location to evaluate the estimate at.}

\item{mc_cores}{Integer: number of cores to use to evaluate the constituent
estimates in parallel. Passed to \code{mclapply}. Suggest leaving at 1
unless each sub estimate is made up of a very large number of samples (more
than 10,000), otherwise the overhead from \code{fork}ing results in a net
slowdown.}
}
\value{
Double: self ratio value
}
\description{
Evaluate a weighted self ratio estimate at a set of numerator and denominator
points
}
\details{
NOTE: This differs from the earlier version, everything is now pointwise due
to the requirement to \code{work} in N-dimensions.
}
