% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate a \code{wsre} object}
\usage{
evaluate(wsre_obj, x_nu, x_de)
}
\arguments{
\item{wsre_obj}{\code{wsre} object: From \code{\link{wsre}}}

\item{x_nu}{Numeric Vector: vector of length one or more; numerator 
location to evaluate the estimate at.}

\item{x_de}{Numeric Vector: vector of length one or more; denominator 
location to evaluate the estimate at.}
}
\value{
Double: self ratio value
}
\description{
Evaluate a weighted self ratio estimate at a set of numerator and denominator
points
}
\details{
NOTE: This differs from the earlier version, everything is now pointwise due
to the requirement to `work` in N-dimensions.
}
