% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{wsre}
\alias{wsre}
\title{Weighted Self Ratio Estimation}
\usage{
wsre(
  model_name = names(stanmodels),
  stanmodel = .named_model(model_name),
  stan_data = list(),
  wf_mean = list(-3, 3, 5),
  wf_pars = list(wf_sd = as.array(c(2)), wf_exponent = 1, target_dimension = 1),
  n_mcmc_samples = 1200,
  stan_control_params = list(adapt_delta = 0.98, max_treedepth = 13),
  flog_threshold = futile.logger::INFO,
  mc_cores = 1
)
}
\arguments{
\item{model_name}{String: The name of a \code{.stan} file in the
\code{inss/stan} directory.}

\item{stanmodel}{stanmodel: output from \code{\link[rstan]{stan_model}}.
Must follow the naming convention in the vignette.}

\item{stan_data}{list: other data to pass to Stan. Not related to weighting
function parameters.}

\item{wf_mean}{List: target-dimension vectors of means of weighting
functions.}

\item{wf_pars}{Named List: see default for structure; other parameters for
the weighting function. Must have wf_sd as an array, even if univariate.
Stan will throw array/vector type errors if you get this wrong.}

\item{n_mcmc_samples}{Integer: number of MCMC samples to draw from each of
the targets, after a fixed 2000 iteration warmup.}

\item{stan_control_params}{Named List: see \code{control} section of
\code{\link[rstan]{stan}}.}

\item{flog_threshold}{Special: See
\code{\link[futile.logger]{futile.logger-package}}. Set to
\code{futile.logger::TRACE} to see informational messages about what is going
on.}
}
\value{
A \code{wsre} object for saving / use in other functions
}
\description{
Estimates the self ratio of a given model using a number of weighting
functions.
}
